requirements.txt
selenium
webdriver-manager
ollama
python-dotenv
